backend/src/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.3
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.1.2
Pygments==2.6.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
